# perf(runtime/spawn): collect output using `op_read_all`