@marcosc90 2022-11-13
> @marcosc90 Can you drop the benchmark source somewhere in `cli/bench`?

Done @littledivy