@marcosc90 2022-11-14
`command_test.ts` are flaky (or broken) in `main`

https://github.com/denoland/deno/actions/runs/3457193770/jobs/5770460515#step:34:3397